controllers and converters created

+++ src/main/java/pl/shellchuck/charity/repository/InstitutionRepository.java
@@ -3,5 +3,5 @@ package pl.shellchuck.charity.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.shellchuck.charity.entity.Institution;
 
-public interface InstitutionRepository extends JpaRepository<Institution, Integer> {
+public interface InstitutionRepository extends JpaRepository<Institution, Long> {
 }
